[provisioning] move fuse lib generator to crate and build script
This fixes #948. In doing so, we remove the legacy xtask flow.
